+++ lib/screens/food_logging.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:dropdown_button2/dropdown_button2.dart';
-import 'package:cloud_firestore/cloud_firestore.dart';
+//import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:url_launcher/url_launcher.dart'; // FatSecret snippet code launch
 import 'dart:async'; // for using Timer
 // Packages for handling food information through the server.py file
@@ -303,7 +303,9 @@ class _FoodLoggingState extends State<FoodLogging> {
                     checkTimer
                         ?.cancel(); // Cancel any previous timers to avoid calling the function continuously
                     checkTimer = Timer(
-                      Duration(milliseconds: 500), // Set a timer for 500ms
+                      Duration(
+                        milliseconds: 750,
+                      ), // Set a timer so api calls don't happen too often
                       () {
                         handleApiCall(
                           lastInput,